fix: rename to profile pic
