+++ components/Home/Contacts/SearchBox.tsx
@@ -7,8 +7,8 @@ import useDebounce from '@/hooks/useDebounce';
 
 import { getSearchedChatsAPI } from '@/axios/chatRequest';
 
-import Avatar from '@/components/global/Avatar';
 import LoaderSpinner from '@/components/global/loader/LoaderSpinner';
+import ProfilePic from '@/components/global/ProfilePic';
 import { addToChatBoxList } from '@/redux/slices/messageSlice';
 
 interface SearchBoxProps {
@@ -65,7 +65,7 @@ const SearchBox = ({ setSearchOpen }: SearchBoxProps) => {
                     key={user._id}
                     className="flex cursor-pointer items-center space-x-2 p-2"
                   >
-                    <Avatar width={30} height={30} profileImage={user.profileImage} />
+                    <ProfilePic width={30} height={30} profileImage={user.profileImage} />
                     <span className="ml-3">{user.name}</span>
                   </div>
                 ))}
