+++ components/Home/Feed/InputBoxModal.tsx
@@ -10,9 +10,9 @@ import useNotify from '@/hooks/useNotify';
 
 import { postNewPostAPI } from '@/axios/postRequest';
 
-import Avatar from '@/components/global/Avatar';
 import Loader from '@/components/global/loader/LoaderBounce';
 import LoaderSpinner from '@/components/global/loader/LoaderSpinner';
+import ProfilePic from '@/components/global/ProfilePic';
 import { setNotification } from '@/redux/slices/globalSlice';
 import { setImagesToPost, setPostInputBoxOpen } from '@/redux/slices/postSlice';
 
@@ -96,7 +96,7 @@ const InputBoxModal = () => {
       </div>
       <div className="p-3">
         <div className="flex items-center py-2">
-          <Avatar
+          <ProfilePic
             width={50}
             height={50}
             profileImage={userInfo.profileImage}
