+++ components/Notifications/FriendNotification.tsx
@@ -7,7 +7,7 @@ import { useRouter } from 'next/router';
 import { postFriendRequestAPI, rejectFriendRequestAPI } from '@/axios/friendRequest';
 import { postReadSingleNotificationAPI } from '@/axios/notificationRequest';
 
-import Avatar from '@/components/global/Avatar';
+import Avatar from '@/components/global/ProfilePic';
 import { timeDiff } from '@/lib/dayjs';
 
 interface FriendNotificationProps {
